Remove gutter signs
Errors are shown in the editor, and a summary of all the errors are
shown in the lualine; I don't think there's a need to also show them in
the gutter, which always makes the gutter fatter and takes up space,
even when there are no errors. We have hotkeys to jump from one error to
the next if it's not already clear (which it almost always is from the
editor error) exactly which line an error is for.