Hide gitx by removing its x alias
We're trying to switch to using lazygit full time and having this alias
lets us accidentally get away with using gitx. It's just too easy. It's
instinct to type x. This should stop us when we instinctively reach for
gitx with the `x` command while allowing us to still access it with
`gitx` if we find we really need it over lazygit.