Ignore user-added dotfiles
Every file and folder in the dotfiles/ directory is symbolically linked
to the ~/ (home) directory by the ./install.sh script. Because of that,
user added dotfiles appear in this repository as I use my machine after
installing the dotfiles, causing this git repository to indicate that
the working tree is not clean (because it has untracked files). To
prevent that "dirty" state, we ignore every file and folder in the
dotfiles/ directory, with the exception of the files and folders which
we excplicitly want to be in there.