Update lazygit config
Lazy loaded plugins are not run until either one of their cmd's or key's
are executed. We never use any of these cmd's; we only use the key
shortcut, so we can remove them. Also, if either of the cmd or key table
keys are set, lazy is implicitly true, so we can remove that table key.