Use vim-test to determine test commands
We have keymaps to run tests in vim

* <leader>s runs the test the cursor is over
* <leader>t runs all the tests in the current file
* <leader>a runs all the tests in the suite

Previously, we had a lot of lua logic to determine whether or not the
file was a test file, and if so, what the command to run the test was.

That was hard to maintain because we not only needed to write and
maintain the logic for the three test types we supported (rspec, exunit,
and jest), but also write additional support for new test types. I
recently started a new project at work and someone decided to use
vitest for that project. When I tried to run tests, it just didn't work
because we hadn't written any support for vitest.

Instead of writing our own adapters, we use vim-test to determine the
test command to run, which supports a ton of test types for us.

https://github.com/vim-test/vim-test

It's great.

We did, however, define our own "strategy" for running the tests.
vim-test lets you hook into their api and define a custom "strategy",
which determines how the test command is run. We wrote a thing which
sends the test command to the first available tmux pane that's open in
the same directory as the vim buffer that generated the test command.

There are existing options for doing that sort of thing that already
exist, but I have problems with both of them.

There's tslime:

https://github.com/jgdavey/tslime.vim

I used tslime for about a decade. It works. Also, it hasn't been updated
in 5 years. Also, when you run a test, it doesn't just automatically
find the first available tmux pane and run the test command there; it
asks you for the index of the pane you'd like to run the tmux command
in. For years, I tried to always say "0" when asked (it only asks the
first time you try to run a test, and subsequently always uses that pane
index). However, if you say use pane 0, and then swap panes around,
things get messy. If you forget that you're saving that pane at index 0
for running tests and open another vim instance there, tslime will
paste the test-running command into your vim buffer. Our strategy is,
in my opinion, a slight improvement over this. It's very little code
and it will never prompt for a pane index or paste into a vim buffer by
accident. If no suitable tmux pane is available, it'll notify you.

There's also Slimux:

https://github.com/esamattis/slimux

Which, coincidentally, also hasn't been updated in 5 years. Also,
there's a notice at the top of the README saying it's no longer
being maintained.