Avoid errors when closing last pane
Sometimes vim tries to run close instead of q when closing the last
pane. I'm not sure what causes it; but we'll always use q here now since
we're not trying to do anything special with the q keymap anymore, which
had been causing problems before and resulted in the needed for calling
close v q on a case by case basis earlier. That stuff is gone.