Revert "Remove location from lualine"
This reverts commit dd774ca86be4df6ce5cb22110fa6f723c1f73325.

It turns out, we do reference the column number. Prettier formats our
typescript code and it usually has an 80 character line length limit.
Sometimes, commands that can be kept on a single line are entered with
newlines in that, in which case, prettier lets you split the command
across multiple lines even though it could fit on a single line.
Sometimes we want to keep a command on a single line, so it's helpful to
know whether it's 80 characters or not.