Install yarn with asdf instead of homebrew
We're working on a project temporarily which requires a different
version of yarn than we use. Our version is installed with homebrew,
so we don't have yarn versioning options. This commit removes the
homebrew installed version of yarn from the dotfiles and instead uses
homebrew to install the things that asdf will need to install yarn
when running:

```
asdf plugin add yarn
```