Write ~/.gitconfig in ./install.sh
Previously, in order to make these dotfiles less specific to me and
usable for others, we setup entering in your git name and email during
install so that my name and email wasn't hard-set in the gitconfig.

We used a template file to generate the gitconfig. After installing the
dotfiles and being prompted to enter your git name and email, the
repository's local status was always dirty because of the automated
edit to the template gitconfig file. I don't want my git repo's status
to be dirty. Now we gitignore the gitconfig file and write the entire
file during installation.