Update deps and prevent new mason defaults
Update dependencys and prevent mason from starting LSPs with default
configurations given we set them each up with a custom configuration.

We noticed that this was a problem after updating the dependencies. When
we edited LUA files, two instances of the lua_ls language server
attached; one with our custom config, which tells the lsp not to
complain about an undefined "vim" global, and one with a default config
which instructs the lsp to continue to complain about undefined 'vim'
globals.

This commit should prevent multiple LSPs - one with the wrong (default)
config and one with the correct, custom config from attaching to each
buffer.