Collapse nvim tree when closed
Sometimes, when I've had a project open for a while, I realize all the
folders in the nvim tree are open and it's more confusing than helpful.
This change ensures that when it's opened (to the file I'm editing) that
all the other folders for other files that I've opened it for are not
still open.