Improve ZSH prompt
Our zsh prompt says the name of the branch we're on when we're in a
directory that's a git repository.

When we're in a detached HEAD state, for years, it has indicated the
branch name as "(HEAD" with just that one opening paren. This commit
fixes that indicator to instead say just "HEAD".